+++ 727 server/src/com/rs/game/dialogue/DialogueOptionEvent.java
@@ -4,31 +4,31 @@ public class DialogueOptionEvent extends DialogueEvent{
 	
 	
 	public DialogueOptionEvent(String title, String name1, Runnable task1){
-		super(3, title);
+		super((byte) 3, title);
 		this.names = new String[] { name1 };
 		this.tasks = new Runnable[] { task1 };
 	}
 	
 	public DialogueOptionEvent(String title, String name1, Runnable task1, String name2, Runnable task2){
-		super(3, title);
+		super((byte) 3, title);
 		this.names = new String[] { name1, name2 };
 		this.tasks = new Runnable[] { task1, task2 };
 	}
 	
 	public DialogueOptionEvent(String title, String name1, Runnable task1, String name2, Runnable task2, String name3, Runnable task3){
-		super(3, title);
+		super((byte) 3, title);
 		this.names = new String[] { name1, name2, name3 };
 		this.tasks = new Runnable[] { task1, task2, task3};
 	}
 
 	public DialogueOptionEvent(String title, String name1, Runnable task1, String name2, Runnable task2, String name3, Runnable task3, String name4, Runnable task4){
-		super(3, title);
+		super((byte) 3, title);
 		this.names = new String[] { name1, name2, name3, name4 };
 		this.tasks = new Runnable[] { task1, task2, task3, task4 };
 	}
 	
 	public DialogueOptionEvent(String title, String name1, Runnable task1, String name2, Runnable task2, String name3, Runnable task3, String name4, Runnable task4, String name5, Runnable task5){
-		super(3, title);
+		super((byte) 3, title);
 		this.names = new String[] { name1, name2, name3, name4, name5 };
 		this.tasks = new Runnable[] { task1, task2, task3, task4, task5 };
 	}
